refactor: 로직을 별도 파일로 분리
- 컴포넌트 내부에서는 가능하면 선언적으로 동작을 표현하는 데에 초점을 맞추어 리팩토링 함
- 데이터의 변경이 컴포넌트 내부에서는 드러나지 않도록 작성함
- 각 컴포넌트에서 사용되는 로직을 각각 별도 파일로 분리함
- 분리한 로직은 page컴포넌트에서 actions객체로 묶어서 import함
- page컴포넌트는 actions객체와 updater function(setState)을 props를 통해 하위 컴포넌트에 전달
- 컨테이너 컴포넌트에서 actions객체와 updater function을 이용해 핸들러 함수를 작성
- 로직이 변경되더라도 전달과정은 신경쓸 필요가 없도록 actions이라는 단일 prop으로 전달하도록 함
- 로직이 변경되면 로직을 정의한 파일과 호출하는 파일만 수정하면 됨
- 수정 위치를 파악하기 쉽도록 로직을 각 컴포넌트 별로 분리함
- 비동기로 동작하는 saveImage()와 addImage() 함수가 이전 state를 인자로 받지 않도록 수정함
  - state를 인자로 받지 않으면 해당 컴포넌트에서 state를 prop으로 전달받지 않을 수 있음
