+++ src/components/PostUploader/ImageUploader/ImageEditor/ImageEditor.jsx
@@ -1,11 +1,10 @@
-import React, { Suspense, lazy } from 'react';
+import React, { useState, Suspense, lazy } from 'react';
 import classNames from 'classnames/bind';
 import styles from './ImageEditor.scss';
 import CommonPost from '../../../CommonPost/CommonPost';
 import EditorHeader from './EditorHeader';
 import EditorFooter from './EditorFooter';
 import 'cropperjs/dist/cropper.css';
-import { useState } from 'react';
 import { asyncPipe } from '../../../../utils/utils';
 
 const cx = classNames.bind(styles);
@@ -14,7 +13,14 @@ const Cropper = lazy(() =>
   import(/* webpackChunkName: "cropper" */ 'react-cropper')
 );
 
-const ImageEditor = ({ setImages, images, src, targetIndex, onClose }) => {
+const ImageEditor = ({
+  setImages,
+  actions,
+  src,
+  originalFile,
+  targetIndex,
+  onClose
+}) => {
   const [cropper, setCropper] = useState(null);
   const ref = cropper => setCropper(cropper);
 
@@ -23,7 +29,7 @@ const ImageEditor = ({ setImages, images, src, targetIndex, onClose }) => {
 
   const saveImage = () =>
     asyncPipe(
-      () => UPDATE_IMAGE(images, setImages, cropper, targetIndex),
+      () => actions.UPDATE_IMAGE(setImages, cropper, targetIndex, originalFile),
       onClose
     )();
 
@@ -54,48 +60,3 @@ const ImageEditor = ({ setImages, images, src, targetIndex, onClose }) => {
 };
 
 export default ImageEditor;
-
-const CANVAS_OPTIONS = {
-  maxWidth: 4096,
-  maxHeight: 4096,
-  fillColor: '#fff',
-  imageSmoothingEnabled: false,
-  imageSmoothingQuality: 'high'
-};
-
-const UPDATE_IMAGE = (images, setImages, cropper, index) => {
-  asyncPipe(
-    () => cropper.getCroppedCanvas(CANVAS_OPTIONS), //sync function
-    canvas => convertCanvasToFile(canvas, images[index].original),
-    file => updateImage(images, index, file),
-    updatedImages => setImages(updatedImages)
-  )();
-};
-
-const convertCanvasToFile = (canvasElement, originalFile) => {
-  const { name, type } = originalFile;
-  return new Promise(resolve => {
-    canvasElement.toBlob(blob => {
-      const file = new File([blob], name, { type });
-      resolve(file);
-    }, type);
-  });
-};
-
-const updateImage = async (images, index, forUpload) => {
-  const previewURL = await readFileAsDataURL(forUpload);
-  return updateArrayWithObject(images, index, { forUpload, previewURL });
-};
-
-const updateArrayWithObject = (images, index, item) =>
-  images.map((image, idx) => (idx === index ? { ...image, ...item } : image));
-
-const readFileAsDataURL = file => {
-  return new Promise(resolve => {
-    const reader = new FileReader();
-    reader.addEventListener('load', ({ target }) => {
-      resolve(target.result);
-    });
-    reader.readAsDataURL(file);
-  });
-};
