+++ src/pages/PostUploadPage/actions.js
@@ -1,0 +1,8 @@
+import { UPDATE_IMAGE } from '../../components/PostUploader/ImageUploader/ImageEditor/ImageEditorLogic';
+import {
+  ADD_IMAGES,
+  DELETE_IMAGE,
+  UPDATE_REPRESENTATIVE
+} from '../../components/PostUploader/ImageUploader/ImageUploaderLogic';
+
+export { ADD_IMAGES, DELETE_IMAGE, UPDATE_REPRESENTATIVE, UPDATE_IMAGE };
